requirements.txt
Django==1.7.4
decorator==3.4.0
dj-database-url==0.2.2
django-pylibmc==0.5.0
dj-static==0.0.6
django-nested-inline==0.3.3
gunicorn==19.1.1
inflection==0.2.1
pgwrap==0.7
psycopg2==2.5.4
pylibmc==1.3.0
python-dateutil==1.5
pytz==2013b0
requests==2.2.1
six==1.4.1
static==0.4
wsgiref==0.1.2
pillow==2.7.0
